Create codeinc++
The hospital management system program in C++ that I provided earlier uses classes and methods to represent patients and the hospital itself, and it allows users to add and display patients in the system. Here's a more detailed explanation of how the program works.

he Patient Class
The Patient class represents a patient in the hospital. It has properties like name, age, address, phone, email, gender, blood group, disease, doctor, and date. In C++, we define a class by using the class keyword, followed by the class name and the class body enclosed in curly braces. 
The public: keyword specifies that all the properties and methods defined below it are accessible outside of the class. The string data type is used to represent strings of characters, and the int data type is used to represent integers.

The Patient class also has a constructor, which is a special method that gets called when a new object of the class is created. The constructor takes 10 parameters, one for each property of the Patient class, and initializes them using the this keyword. The this keyword is used to refer to the current object, and it's used here to set the object's properties equal to the constructor parameters.

The Hospital Class
The Hospital class represents the hospital itself. It has a vector of Patient objects to keep track of all patients. In C++, we define a class by using the class keyword, followed by the class name and the class body enclosed in curly braces.
The public: keyword specifies that all the properties and methods defined below it are accessible outside of the class. The vector data type is used to represent a dynamic array of objects. Here, we're using a vector of Patient objects to represent the patients in